+++ include/crm/pengine/Makefile.am
@@ -22,4 +22,4 @@ MAINTAINERCLEANFILES    = Makefile.in
 headerdir=$(pkgincludedir)/crm/pengine
 
 noinst_HEADERS		= internal.h rules_internal.h
-header_HEADERS		= common.h complex.h remote.h rules.h status.h
+header_HEADERS		= common.h complex.h remote.h rules.h status.h print_ml_status.h
