+++ maint/bumplibs.sh
@@ -15,8 +15,8 @@ headers[crmservice]="include/crm/services.h"
 headers[transitioner]="include/crm/transition.h"
 headers[cib]="include/crm/cib.h include/crm/cib/util.h"
 headers[pe_rules]="include/crm/pengine/rules.h"
-headers[pe_status]="include/crm/pengine/common.h  include/crm/pengine/complex.h  include/crm/pengine/rules.h  include/crm/pengine/status.h"
-headers[pengine]="include/crm/pengine/common.h  include/crm/pengine/complex.h  include/crm/pengine/rules.h  include/crm/pengine/status.h"
+headers[pe_status]="include/crm/pengine/common.h  include/crm/pengine/complex.h  include/crm/pengine/rules.h  include/crm/pengine/status.h include/crm/pengine/print_ml_status.h"
+headers[pengine]="include/crm/pengine/common.h  include/crm/pengine/complex.h  include/crm/pengine/rules.h  include/crm/pengine/status.h include/crm/pengine/print_ml_status.h"
 headers[stonithd]="include/crm/stonith-ng.h"
 headers[lrmd]="include/crm/lrmd.h"
