+++ tools/crm_mon.c
@@ -35,6 +35,7 @@
 
 #include <crm/cib/internal.h>
 #include <crm/pengine/status.h>
+#include <crm/pengine/print_ml_status.h>
 #include <crm/pengine/internal.h>
 #include <../lib/pengine/unpack.h>
 #include <pacemaker-internal.h>
@@ -48,64 +49,22 @@ static int cib_connect(gboolean full);
 static void mon_st_callback_event(stonith_t * st, stonith_event_t * e);
 static void mon_st_callback_display(stonith_t * st, stonith_event_t * e);
 static void kick_refresh(gboolean data_updated);
-static char *get_node_display_name(node_t *node);
-
-/*
- * Definitions indicating which items to print
- */
-
-#define mon_show_times         (0x0001U)
-#define mon_show_stack         (0x0002U)
-#define mon_show_dc            (0x0004U)
-#define mon_show_count         (0x0008U)
-#define mon_show_nodes         (0x0010U)
-#define mon_show_resources     (0x0020U)
-#define mon_show_attributes    (0x0040U)
-#define mon_show_failcounts    (0x0080U)
-#define mon_show_operations    (0x0100U)
-#define mon_show_tickets       (0x0200U)
-#define mon_show_bans          (0x0400U)
-#define mon_show_fence_history (0x0800U)
-
-#define mon_show_headers       (mon_show_times | mon_show_stack | mon_show_dc \
-                               | mon_show_count)
-#define mon_show_default       (mon_show_headers | mon_show_nodes \
-                               | mon_show_resources)
-#define mon_show_all           (mon_show_default | mon_show_attributes \
-                               | mon_show_failcounts | mon_show_operations \
-                               | mon_show_tickets | mon_show_bans \
-                               | mon_show_fence_history)
-
-static unsigned int show = mon_show_default;
 
 /*
  * Definitions indicating how to output
  */
-
-enum mon_output_format_e {
-    mon_output_none,
-    mon_output_monitor,
-    mon_output_plain,
-    mon_output_console,
-    mon_output_xml,
-    mon_output_html,
-    mon_output_cgi
-} output_format = mon_output_console;
 
 static char *output_filename = NULL;   /* if sending output to a file, its name */
 
 /* other globals */
 static char *pid_file = NULL;
 
-static gboolean group_by_node = FALSE;
-static gboolean inactive_resources = FALSE;
 static int reconnect_msec = 5000;
 static gboolean daemonize = FALSE;
 static GMainLoop *mainloop = NULL;
 static guint timer_id = 0;
 static mainloop_timer_t *refresh_timer = NULL;
 static pe_working_set_t *mon_data_set = NULL;
-static GList *attr_list = NULL;
 
 static const char *external_agent = NULL;
 static const char *external_recipient = NULL;
@@ -116,26 +75,13 @@ static xmlNode *current_cib = NULL;
 
 static gboolean one_shot = FALSE;
 static gboolean has_warnings = FALSE;
-static gboolean print_timing = FALSE;
 static gboolean watch_fencing = FALSE;
-static gboolean fence_history = FALSE;
-static gboolean fence_full_history = FALSE;
+static struct print_params_t pp = { mon_show_default, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, "" };
 static gboolean fence_connect = FALSE;
 static int fence_history_level = 1;
-static gboolean print_brief = FALSE;
-static gboolean print_pending = TRUE;
-static gboolean print_clone_detail = FALSE;
 #if CURSES_ENABLED
 static gboolean curses_console_initialized = FALSE;
 #endif
-
-/* FIXME allow, detect, and correctly interpret glob pattern or regex? */
-const char *print_neg_location_prefix = "";
-
-/* Never display node attributes whose name starts with one of these prefixes */
-#define FILTER_STR { CRM_FAIL_COUNT_PREFIX, CRM_LAST_FAILURE_PREFIX,       \
-                     "shutdown", "terminate", "standby", "probe_complete", \
-                     "#", NULL }
 
 long last_refresh = 0;
 crm_trigger_t *refresh_trigger = NULL;
@@ -147,33 +93,6 @@ crm_trigger_t *refresh_trigger = NULL;
 #define MON_STATUS_WARN    CRM_EX_ERROR
 #define MON_STATUS_CRIT    CRM_EX_INVALID_PARAM
 #define MON_STATUS_UNKNOWN CRM_EX_UNIMPLEMENT_FEATURE
-
-/* Convenience macro for prettifying output (e.g. "node" vs "nodes") */
-#define s_if_plural(i) (((i) == 1)? "" : "s")
-
-#if CURSES_ENABLED
-#  define print_dot() if (output_format == mon_output_console) { \
-	printw(".");				\
-	clrtoeol();				\
-	refresh();				\
-    } else {					\
-	fprintf(stdout, ".");			\
-    }
-#else
-#  define print_dot() fprintf(stdout, ".");
-#endif
-
-#if CURSES_ENABLED
-#  define print_as(fmt, args...) if (output_format == mon_output_console) { \
-	printw(fmt, ##args);				\
-	clrtoeol();					\
-	refresh();					\
-    } else {						\
-	fprintf(stdout, fmt, ##args);			\
-    }
-#else
-#  define print_as(fmt, args...) fprintf(stdout, fmt, ##args);
-#endif
 
 static void
 blank_screen(void)
@@ -482,60 +401,60 @@ detect_user_input(GIOChannel *channel, GIOCondition condition, gpointer unused)
         switch (c) {
             case 'm':
                 if (!fence_history_level) {
-                    fence_history = TRUE;
+                    pp.fence_history = TRUE;
                     fence_connect = TRUE;
                     if (st == NULL) {
                         mon_cib_connection_destroy(NULL);
                     }
                 }
-                show ^= mon_show_fence_history;
+                pp.show ^= mon_show_fence_history;
                 break;
             case 'c':
-                show ^= mon_show_tickets;
+                pp.show ^= mon_show_tickets;
                 break;
             case 'f':
-                show ^= mon_show_failcounts;
+                pp.show ^= mon_show_failcounts;
                 break;
             case 'n':
-                group_by_node = ! group_by_node;
+                pp.group_by_node = ! pp.group_by_node;
                 break;
             case 'o':
-                show ^= mon_show_operations;
-                if ((show & mon_show_operations) == 0) {
-                    print_timing = 0;
+                pp.show ^= mon_show_operations;
+                if ((pp.show & mon_show_operations) == 0) {
+                    pp.print_timing = 0;
                 }
                 break;
             case 'r':
-                inactive_resources = ! inactive_resources;
+                pp.inactive_resources = ! pp.inactive_resources;
                 break;
             case 'R':
-                print_clone_detail = ! print_clone_detail;
+                pp.print_clone_detail = ! pp.print_clone_detail;
                 break;
             case 't':
-                print_timing = ! print_timing;
-                if (print_timing) {
-                    show |= mon_show_operations;
+                pp.print_timing = ! pp.print_timing;
+                if (pp.print_timing) {
+                    pp.show |= mon_show_operations;
                 }
                 break;
             case 'A':
-                show ^= mon_show_attributes;
+                pp.show ^= mon_show_attributes;
                 break;
             case 'L':
-                show ^= mon_show_bans;
+                pp.show ^= mon_show_bans;
                 break;
             case 'D':
                 /* If any header is shown, clear them all, otherwise set them all */
-                if (show & mon_show_headers) {
-                    show &= ~mon_show_headers;
+                if (pp.show & mon_show_headers) {
+                    pp.show &= ~mon_show_headers;
                 } else {
-                    show |= mon_show_headers;
+                    pp.show |= mon_show_headers;
                 }
                 break;
             case 'b':
-                print_brief = ! print_brief;
+                pp.print_brief = ! pp.print_brief;
                 break;
             case 'j':
-                print_pending = ! print_pending;
+                pp.print_pending = ! pp.print_pending;
                 break;
             case '?':
                 config_mode = TRUE;
@@ -551,19 +470,19 @@ detect_user_input(GIOChannel *channel, GIOCondition condition, gpointer unused)
 
         print_as("Display option change mode\n");
         print_as("\n");
-        print_option_help('c', show & mon_show_tickets);
-        print_option_help('f', show & mon_show_failcounts);
-        print_option_help('n', group_by_node);
-        print_option_help('o', show & mon_show_operations);
-        print_option_help('r', inactive_resources);
-        print_option_help('t', print_timing);
-        print_option_help('A', show & mon_show_attributes);
-        print_option_help('L', show & mon_show_bans);
-        print_option_help('D', (show & mon_show_headers) == 0);
-        print_option_help('R', print_clone_detail);
-        print_option_help('b', print_brief);
-        print_option_help('j', print_pending);
-        print_option_help('m', (show & mon_show_fence_history));
+        print_option_help('c', pp.show & mon_show_tickets);
+        print_option_help('f', pp.show & mon_show_failcounts);
+        print_option_help('n', pp.group_by_node);
+        print_option_help('o', pp.show & mon_show_operations);
+        print_option_help('r', pp.inactive_resources);
+        print_option_help('t', pp.print_timing);
+        print_option_help('A', pp.show & mon_show_attributes);
+        print_option_help('L', pp.show & mon_show_bans);
+        print_option_help('D', (pp.show & mon_show_headers) == 0);
+        print_option_help('R', pp.print_clone_detail);
+        print_option_help('b', pp.print_brief);
+        print_option_help('j', pp.print_pending);
+        print_option_help('m', (pp.show & mon_show_fence_history));
         print_as("\n");
         print_as("Toggle fields via field letter, type any other key to return");
     }
@@ -614,16 +533,16 @@ main(int argc, char **argv)
                 crm_bump_log_level(argc, argv);
                 break;
             case 'Q':
-                show &= ~mon_show_times;
+                pp.show &= ~mon_show_times;
                 break;
             case 'i':
                 reconnect_msec = crm_get_msec(optarg);
                 break;
             case 'n':
-                group_by_node = TRUE;
+                pp.group_by_node = TRUE;
                 break;
             case 'r':
-                inactive_resources = TRUE;
+                pp.inactive_resources = TRUE;
                 break;
             case 'W':
                 watch_fencing = TRUE;
@@ -636,36 +555,36 @@ main(int argc, char **argv)
                 daemonize = TRUE;
                 break;
             case 't':
-                print_timing = TRUE;
-                show |= mon_show_operations;
+                pp.print_timing = TRUE;
+                pp.show |= mon_show_operations;
                 break;
             case 'o':
-                show |= mon_show_operations;
+                pp.show |= mon_show_operations;
                 break;
             case 'f':
-                show |= mon_show_failcounts;
+                pp.show |= mon_show_failcounts;
                 break;
             case 'A':
-                show |= mon_show_attributes;
+                pp.show |= mon_show_attributes;
                 break;
             case 'L':
-                show |= mon_show_bans;
-                print_neg_location_prefix = optarg? optarg : "";
+                pp.show |= mon_show_bans;
+                pp.print_neg_location_prefix = optarg? optarg : "";
                 break;
             case 'D':
-                show &= ~mon_show_headers;
+                pp.show &= ~mon_show_headers;
                 break;
             case 'b':
-                print_brief = TRUE;
+                pp.print_brief = TRUE;
                 break;
             case 'j':
-                print_pending = TRUE;
+                pp.print_pending = TRUE;
                 break;
             case 'R':
-                print_clone_detail = TRUE;
+                pp.print_clone_detail = TRUE;
                 break;
             case 'c':
-                show |= mon_show_tickets;
+                pp.show |= mon_show_tickets;
                 break;
             case 'p':
                 free(pid_file);
@@ -778,13 +697,13 @@ main(int argc, char **argv)
 
     switch (fence_history_level) {
         case 3:
-            fence_full_history = TRUE;
+            pp.fence_full_history = TRUE;
             /* fall through to next lower level */
         case 2:
-            show |= mon_show_fence_history;
+            pp.show |= mon_show_fence_history;
             /* fall through to next lower level */
         case 1:
-            fence_history = TRUE;
+            pp.fence_history = TRUE;
             fence_connect = TRUE;
             break;
         default:
@@ -812,8 +731,8 @@ main(int argc, char **argv)
 
     /* XML output always prints everything */
     if (output_format == mon_output_xml) {
-        show = mon_show_all;
-        print_timing = TRUE;
+        pp.show = mon_show_all;
+        pp.print_timing = TRUE;
     }
 
     if (one_shot) {
@@ -952,26 +871,6 @@ main(int argc, char **argv)
 	has_warnings = TRUE;			\
     } while(0)
 
-static int
-count_resources(pe_working_set_t * data_set, resource_t * rsc)
-{
-    int count = 0;
-    GListPtr gIter = NULL;
-
-    if (rsc == NULL) {
-        gIter = data_set->resources;
-    } else if (rsc->children) {
-        gIter = rsc->children;
-    } else {
-        return is_not_set(rsc->flags, pe_rsc_orphan);
-    }
-
-    for (; gIter != NULL; gIter = gIter->next) {
-        count += count_resources(data_set, gIter->data);
-    }
-    return count;
-}
-
 /*!
  * \internal
  * \brief Print one-line status suitable for use with monitoring software
@@ -1025,1971 +924,11 @@ print_simple_status(pe_working_set_t * data_set,
     print_as("\n");
 }
 
-/*!
- * \internal
- * \brief Print a [name]=[value][units] pair, optionally using time string
- *
- * \param[in] stream      File stream to display output to
- * \param[in] name        Name to display
- * \param[in] value       Value to display (or NULL to convert time instead)
- * \param[in] units       Units to display (or NULL for no units)
- * \param[in] epoch_time  Epoch time to convert if value is NULL
- */
-static void
-print_nvpair(FILE *stream, const char *name, const char *value,
-             const char *units, time_t epoch_time)
-{
-    /* print name= */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as(" %s=", name);
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-        case mon_output_xml:
-            fprintf(stream, " %s=", name);
-            break;
-
-        default:
-            break;
-    }
-
-    /* If we have a value (and optionally units), print it */
-    if (value) {
-        switch (output_format) {
-            case mon_output_plain:
-            case mon_output_console:
-                print_as("%s%s", value, (units? units : ""));
-                break;
-
-            case mon_output_html:
-            case mon_output_cgi:
-                fprintf(stream, "%s%s", value, (units? units : ""));
-                break;
-
-            case mon_output_xml:
-                fprintf(stream, "\"%s%s\"", value, (units? units : ""));
-                break;
-
-            default:
-                break;
-        }
-
-    /* Otherwise print user-friendly time string */
-    } else {
-        static char empty_str[] = "";
-        char *c, *date_str = asctime(localtime(&epoch_time));
-
-        for (c = (date_str != NULL) ? date_str : empty_str; *c != '\0'; ++c) {
-            if (*c == '\n') {
-                *c = '\0';
-                break;
-            }
-        }
-        switch (output_format) {
-            case mon_output_plain:
-            case mon_output_console:
-                print_as("'%s'", date_str);
-                break;
-
-            case mon_output_html:
-            case mon_output_cgi:
-            case mon_output_xml:
-                fprintf(stream, "\"%s\"", date_str);
-                break;
-
-            default:
-                break;
-        }
-    }
-}
-
-/*!
- * \internal
- * \brief Print whatever is needed to start a node section
- *
- * \param[in] stream     File stream to display output to
- * \param[in] node       Node to print
- */
-static void
-print_node_start(FILE *stream, node_t *node)
-{
-    char *node_name;
-
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            node_name = get_node_display_name(node);
-            print_as("* Node %s:\n", node_name);
-            free(node_name);
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            node_name = get_node_display_name(node);
-            fprintf(stream, "  <h3>Node: %s</h3>\n  <ul>\n", node_name);
-            free(node_name);
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "        <node name=\"%s\">\n", node->details->uname);
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print whatever is needed to end a node section
- *
- * \param[in] stream     File stream to display output to
- */
-static void
-print_node_end(FILE *stream)
-{
-    switch (output_format) {
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, "  </ul>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "        </node>\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print resources section heading appropriate to options
- *
- * \param[in] stream      File stream to display output to
- */
-static void
-print_resources_heading(FILE *stream)
-{
-    const char *heading;
-
-    if (group_by_node) {
-
-        /* Active resources have already been printed by node */
-        heading = (inactive_resources? "Inactive resources" : NULL);
-
-    } else if (inactive_resources) {
-        heading = "Full list of resources";
-
-    } else {
-        heading = "Active resources";
-    }
-
-    /* Print section heading */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("\n%s:\n\n", heading);
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " <hr />\n <h2>%s</h2>\n", heading);
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    <resources>\n");
-            break;
-
-        default:
-            break;
-    }
-
-}
-
-/*!
- * \internal
- * \brief Print whatever resource section closing is appropriate
- *
- * \param[in] stream     File stream to display output to
- */
-static void
-print_resources_closing(FILE *stream, gboolean printed_heading)
-{
-    const char *heading;
-
-    /* What type of resources we did or did not display */
-    if (group_by_node) {
-        heading = "inactive ";
-    } else if (inactive_resources) {
-        heading = "";
-    } else {
-        heading = "active ";
-    }
-
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            if (!printed_heading) {
-                print_as("\nNo %sresources\n\n", heading);
-            }
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            if (!printed_heading) {
-                fprintf(stream, " <hr />\n <h2>No %sresources</h2>\n", heading);
-            }
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    %s\n",
-                    (printed_heading? "</resources>" : "<resources/>"));
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print whatever resource section(s) are appropriate
- *
- * \param[in] stream     File stream to display output to
- * \param[in] data_set   Cluster state to display
- * \param[in] print_opts  Bitmask of pe_print_options
- */
-static void
-print_resources(FILE *stream, pe_working_set_t *data_set, int print_opts)
-{
-    GListPtr rsc_iter;
-    const char *prefix = NULL;
-    gboolean printed_heading = FALSE;
-    gboolean brief_output = print_brief;
-
-    /* If we already showed active resources by node, and
-     * we're not showing inactive resources, we have nothing to do
-     */
-    if (group_by_node && !inactive_resources) {
-        return;
-    }
-
-    /* XML uses an indent, and ignores brief option for resources */
-    if (output_format == mon_output_xml) {
-        prefix = "        ";
-        brief_output = FALSE;
-    }
-
-    /* If we haven't already printed resources grouped by node,
-     * and brief output was requested, print resource summary */
-    if (brief_output && !group_by_node) {
-        print_resources_heading(stream);
-        printed_heading = TRUE;
-        print_rscs_brief(data_set->resources, NULL, print_opts, stream,
-                         inactive_resources);
-    }
-
-    /* For each resource, display it if appropriate */
-    for (rsc_iter = data_set->resources; rsc_iter != NULL; rsc_iter = rsc_iter->next) {
-        resource_t *rsc = (resource_t *) rsc_iter->data;
-
-        /* Complex resources may have some sub-resources active and some inactive */
-        gboolean is_active = rsc->fns->active(rsc, TRUE);
-        gboolean partially_active = rsc->fns->active(rsc, FALSE);
-
-        /* Skip inactive orphans (deleted but still in CIB) */
-        if (is_set(rsc->flags, pe_rsc_orphan) && !is_active) {
-            continue;
-
-        /* Skip active resources if we already displayed them by node */
-        } else if (group_by_node) {
-            if (is_active) {
-                continue;
-            }
-
-        /* Skip primitives already counted in a brief summary */
-        } else if (brief_output && (rsc->variant == pe_native)) {
-            continue;
-
-        /* Skip resources that aren't at least partially active,
-         * unless we're displaying inactive resources
-         */
-        } else if (!partially_active && !inactive_resources) {
-            continue;
-        }
-
-        /* Print this resource */
-        if (printed_heading == FALSE) {
-            print_resources_heading(stream);
-            printed_heading = TRUE;
-        }
-        rsc->fns->print(rsc, prefix, print_opts, stream);
-    }
-
-    print_resources_closing(stream, printed_heading);
-}
-
-/*!
- * \internal
- * \brief Print heading for resource history
- *
- * \param[in] stream      File stream to display output to
- * \param[in] data_set    Current state of CIB
- * \param[in] node        Node that ran this resource
- * \param[in] rsc         Resource to print
- * \param[in] rsc_id      ID of resource to print
- * \param[in] all         Whether to print every resource or just failed ones
- */
-static void
-print_rsc_history_start(FILE *stream, pe_working_set_t *data_set, node_t *node,
-                        resource_t *rsc, const char *rsc_id, gboolean all)
-{
-    time_t last_failure = 0;
-    int failcount = rsc?
-                    pe_get_failcount(node, rsc, &last_failure, pe_fc_default,
-                                     NULL, data_set)
-                    : 0;
-
-    if (!all && !failcount && (last_failure <= 0)) {
-        return;
-    }
-
-    /* Print resource ID */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("   %s:", rsc_id);
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, "   <li>%s:", rsc_id);
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "            <resource_history id=\"%s\"", rsc_id);
-            break;
-
-        default:
-            break;
-    }
-
-    /* If resource is an orphan, that's all we can say about it */
-    if (rsc == NULL) {
-        switch (output_format) {
-            case mon_output_plain:
-            case mon_output_console:
-                print_as(" orphan");
-                break;
-
-            case mon_output_html:
-            case mon_output_cgi:
-                fprintf(stream, " orphan");
-                break;
-
-            case mon_output_xml:
-                fprintf(stream, " orphan=\"true\"");
-                break;
-
-            default:
-                break;
-        }
-
-    /* If resource is not an orphan, print some details */
-    } else if (all || failcount || (last_failure > 0)) {
-
-        /* Print migration threshold */
-        switch (output_format) {
-            case mon_output_plain:
-            case mon_output_console:
-                print_as(" migration-threshold=%d", rsc->migration_threshold);
-                break;
-
-            case mon_output_html:
-            case mon_output_cgi:
-                fprintf(stream, " migration-threshold=%d", rsc->migration_threshold);
-                break;
-
-            case mon_output_xml:
-                fprintf(stream, " orphan=\"false\" migration-threshold=\"%d\"",
-                        rsc->migration_threshold);
-                break;
-
-            default:
-                break;
-        }
-
-        /* Print fail count if any */
-        if (failcount > 0) {
-            switch (output_format) {
-                case mon_output_plain:
-                case mon_output_console:
-                    print_as(" " CRM_FAIL_COUNT_PREFIX "=%d", failcount);
-                    break;
-
-                case mon_output_html:
-                case mon_output_cgi:
-                    fprintf(stream, " " CRM_FAIL_COUNT_PREFIX "=%d", failcount);
-                    break;
-
-                case mon_output_xml:
-                    fprintf(stream, " " CRM_FAIL_COUNT_PREFIX "=\"%d\"",
-                            failcount);
-                    break;
-
-                default:
-                    break;
-            }
-        }
-
-        /* Print last failure time if any */
-        if (last_failure > 0) {
-            print_nvpair(stream, CRM_LAST_FAILURE_PREFIX, NULL, NULL,
-                         last_failure);
-        }
-    }
-
-    /* End the heading */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("\n");
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, "\n    <ul>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, ">\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print closing for resource history
- *
- * \param[in] stream      File stream to display output to
- */
-static void
-print_rsc_history_end(FILE *stream)
-{
-    switch (output_format) {
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, "    </ul>\n   </li>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "            </resource_history>\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print operation history
- *
- * \param[in] stream        File stream to display output to
- * \param[in] data_set      Current state of CIB
- * \param[in] node          Node this operation is for
- * \param[in] xml_op        Root of XML tree describing this operation
- * \param[in] task          Task parsed from this operation's XML
- * \param[in] interval_ms_s Interval parsed from this operation's XML
- * \param[in] rc            Return code parsed from this operation's XML
- */
-static void
-print_op_history(FILE *stream, pe_working_set_t *data_set, node_t *node,
-                 xmlNode *xml_op, const char *task, const char *interval_ms_s,
-                 int rc)
-{
-    const char *value = NULL;
-    const char *call = crm_element_value(xml_op, XML_LRM_ATTR_CALLID);
-
-    /* Begin the operation description */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("    + (%s) %s:", call, task);
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, "     <li>(%s) %s:", call, task);
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "                <operation_history call=\"%s\" task=\"%s\"",
-                    call, task);
-            break;
-
-        default:
-            break;
-    }
-
-    /* Add name=value pairs as appropriate */
-    if (interval_ms_s && safe_str_neq(interval_ms_s, "0")) {
-        print_nvpair(stream, "interval", interval_ms_s, "ms", 0);
-    }
-    if (print_timing) {
-        int int_value;
-        const char *attr;
-
-        attr = XML_RSC_OP_LAST_CHANGE;
-        value = crm_element_value(xml_op, attr);
-        if (value) {
-            int_value = crm_parse_int(value, NULL);
-            if (int_value > 0) {
-                print_nvpair(stream, attr, NULL, NULL, int_value);
-            }
-        }
-
-        attr = XML_RSC_OP_LAST_RUN;
-        value = crm_element_value(xml_op, attr);
-        if (value) {
-            int_value = crm_parse_int(value, NULL);
-            if (int_value > 0) {
-                print_nvpair(stream, attr, NULL, NULL, int_value);
-            }
-        }
-
-        attr = XML_RSC_OP_T_EXEC;
-        value = crm_element_value(xml_op, attr);
-        if (value) {
-            print_nvpair(stream, attr, value, "ms", 0);
-        }
-
-        attr = XML_RSC_OP_T_QUEUE;
-        value = crm_element_value(xml_op, attr);
-        if (value) {
-            print_nvpair(stream, attr, value, "ms", 0);
-        }
-    }
-
-    /* End the operation description */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as(" rc=%d (%s)\n", rc, services_ocf_exitcode_str(rc));
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " rc=%d (%s)</li>\n", rc, services_ocf_exitcode_str(rc));
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, " rc=\"%d\" rc_text=\"%s\" />\n", rc, services_ocf_exitcode_str(rc));
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print resource operation/failure history
- *
- * \param[in] stream      File stream to display output to
- * \param[in] data_set    Current state of CIB
- * \param[in] node        Node that ran this resource
- * \param[in] rsc_entry   Root of XML tree describing resource status
- * \param[in] operations  Whether to print operations or just failcounts
- */
-static void
-print_rsc_history(FILE *stream, pe_working_set_t *data_set, node_t *node,
-                  xmlNode *rsc_entry, gboolean operations)
-{
-    GListPtr gIter = NULL;
-    GListPtr op_list = NULL;
-    gboolean printed = FALSE;
-    const char *rsc_id = crm_element_value(rsc_entry, XML_ATTR_ID);
-    resource_t *rsc = pe_find_resource(data_set->resources, rsc_id);
-    xmlNode *rsc_op = NULL;
-
-    /* If we're not showing operations, just print the resource failure summary */
-    if (operations == FALSE) {
-        print_rsc_history_start(stream, data_set, node, rsc, rsc_id, FALSE);
-        print_rsc_history_end(stream);
-        return;
-    }
-
-    /* Create a list of this resource's operations */
-    for (rsc_op = __xml_first_child(rsc_entry); rsc_op != NULL; rsc_op = __xml_next(rsc_op)) {
-        if (crm_str_eq((const char *)rsc_op->name, XML_LRM_TAG_RSC_OP, TRUE)) {
-            op_list = g_list_append(op_list, rsc_op);
-        }
-    }
-    op_list = g_list_sort(op_list, sort_op_by_callid);
-
-    /* Print each operation */
-    for (gIter = op_list; gIter != NULL; gIter = gIter->next) {
-        xmlNode *xml_op = (xmlNode *) gIter->data;
-        const char *task = crm_element_value(xml_op, XML_LRM_ATTR_TASK);
-        const char *interval_ms_s = crm_element_value(xml_op,
-                                                      XML_LRM_ATTR_INTERVAL_MS);
-        const char *op_rc = crm_element_value(xml_op, XML_LRM_ATTR_RC);
-        int rc = crm_parse_int(op_rc, "0");
-
-        /* Display 0-interval monitors as "probe" */
-        if (safe_str_eq(task, CRMD_ACTION_STATUS)
-            && ((interval_ms_s == NULL) || safe_str_eq(interval_ms_s, "0"))) {
-            task = "probe";
-        }
-
-        /* Ignore notifies and some probes */
-        if (safe_str_eq(task, CRMD_ACTION_NOTIFY) || (safe_str_eq(task, "probe") && (rc == 7))) {
-            continue;
-        }
-
-        /* If this is the first printed operation, print heading for resource */
-        if (printed == FALSE) {
-            printed = TRUE;
-            print_rsc_history_start(stream, data_set, node, rsc, rsc_id, TRUE);
-        }
-
-        /* Print the operation */
-        print_op_history(stream, data_set, node, xml_op, task, interval_ms_s,
-                         rc);
-    }
-
-    /* Free the list we created (no need to free the individual items) */
-    g_list_free(op_list);
-
-    /* If we printed anything, close the resource */
-    if (printed) {
-        print_rsc_history_end(stream);
-    }
-}
-
-/*!
- * \internal
- * \brief Print node operation/failure history
- *
- * \param[in] stream      File stream to display output to
- * \param[in] data_set    Current state of CIB
- * \param[in] node_state  Root of XML tree describing node status
- * \param[in] operations  Whether to print operations or just failcounts
- */
-static void
-print_node_history(FILE *stream, pe_working_set_t *data_set,
-                   xmlNode *node_state, gboolean operations)
-{
-    node_t *node = pe_find_node_id(data_set->nodes, ID(node_state));
-    xmlNode *lrm_rsc = NULL;
-    xmlNode *rsc_entry = NULL;
-
-    if (node && node->details && node->details->online) {
-        print_node_start(stream, node);
-
-        lrm_rsc = find_xml_node(node_state, XML_CIB_TAG_LRM, FALSE);
-        lrm_rsc = find_xml_node(lrm_rsc, XML_LRM_TAG_RESOURCES, FALSE);
-
-        /* Print history of each of the node's resources */
-        for (rsc_entry = __xml_first_child(lrm_rsc); rsc_entry != NULL;
-             rsc_entry = __xml_next(rsc_entry)) {
-
-            if (crm_str_eq((const char *)rsc_entry->name, XML_LRM_TAG_RESOURCE, TRUE)) {
-                print_rsc_history(stream, data_set, node, rsc_entry, operations);
-            }
-        }
-
-        print_node_end(stream);
-    }
-}
-
-/*!
- * \internal
- * \brief Print extended information about an attribute if appropriate
- *
- * \param[in] data_set  Working set of CIB state
- *
- * \return TRUE if extended information was printed, FALSE otherwise
- * \note Currently, extended information is only supported for ping/pingd
- *       resources, for which a message will be printed if connectivity is lost
- *       or degraded.
- */
-static gboolean
-print_attr_msg(FILE *stream, node_t * node, GListPtr rsc_list, const char *attrname, const char *attrvalue)
-{
-    GListPtr gIter = NULL;
-
-    for (gIter = rsc_list; gIter != NULL; gIter = gIter->next) {
-        resource_t *rsc = (resource_t *) gIter->data;
-        const char *type = g_hash_table_lookup(rsc->meta, "type");
-
-        if (rsc->children != NULL) {
-            if (print_attr_msg(stream, node, rsc->children, attrname, attrvalue)) {
-                return TRUE;
-            }
-        }
-
-        if (safe_str_eq(type, "ping") || safe_str_eq(type, "pingd")) {
-            const char *name = g_hash_table_lookup(rsc->parameters, "name");
-
-            if (name == NULL) {
-                name = "pingd";
-            }
-
-            /* To identify the resource with the attribute name. */
-            if (safe_str_eq(name, attrname)) {
-                int host_list_num = 0;
-                int expected_score = 0;
-                int value = crm_parse_int(attrvalue, "0");
-                const char *hosts = g_hash_table_lookup(rsc->parameters, "host_list");
-                const char *multiplier = g_hash_table_lookup(rsc->parameters, "multiplier");
-
-                if(hosts) {
-                    char **host_list = g_strsplit(hosts, " ", 0);
-                    host_list_num = g_strv_length(host_list);
-                    g_strfreev(host_list);
-                }
-
-                /* pingd multiplier is the same as the default value. */
-                expected_score = host_list_num * crm_parse_int(multiplier, "1");
-
-                switch (output_format) {
-                    case mon_output_plain:
-                    case mon_output_console:
-                        if (value <= 0) {
-                            print_as("\t: Connectivity is lost");
-                        } else if (value < expected_score) {
-                            print_as("\t: Connectivity is degraded (Expected=%d)", expected_score);
-                        }
-                        break;
-
-                    case mon_output_html:
-                    case mon_output_cgi:
-                        if (value <= 0) {
-                            fprintf(stream, " <b>(connectivity is lost)</b>");
-                        } else if (value < expected_score) {
-                            fprintf(stream, " <b>(connectivity is degraded -- expected %d)</b>",
-                                    expected_score);
-                        }
-                        break;
-
-                    case mon_output_xml:
-                        fprintf(stream, " expected=\"%d\"", expected_score);
-                        break;
-
-                    default:
-                        break;
-                }
-                return TRUE;
-            }
-        }
-    }
-    return FALSE;
-}
-
-static int
-compare_attribute(gconstpointer a, gconstpointer b)
-{
-    int rc;
-
-    rc = strcmp((const char *)a, (const char *)b);
-
-    return rc;
-}
-
-static void
-create_attr_list(gpointer name, gpointer value, gpointer data)
-{
-    int i;
-    const char *filt_str[] = FILTER_STR;
-
-    CRM_CHECK(name != NULL, return);
-
-    /* filtering automatic attributes */
-    for (i = 0; filt_str[i] != NULL; i++) {
-        if (g_str_has_prefix(name, filt_str[i])) {
-            return;
-        }
-    }
-
-    attr_list = g_list_insert_sorted(attr_list, name, compare_attribute);
-}
-
 /* structure for passing multiple user data to g_list_foreach() */
 struct mon_attr_data {
     FILE *stream;
     node_t *node;
 };
-
-static void
-print_node_attribute(gpointer name, gpointer user_data)
-{
-    const char *value = NULL;
-    struct mon_attr_data *data = (struct mon_attr_data *) user_data;
-
-    value = pe_node_attribute_raw(data->node, name);
-
-    /* Print attribute name and value */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("    + %-32s\t: %-10s", (char *)name, value);
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(data->stream, "   <li>%s: %s",
-                    (char *)name, value);
-            break;
-
-        case mon_output_xml:
-            fprintf(data->stream,
-                    "            <attribute name=\"%s\" value=\"%s\"",
-                    (char *)name, value);
-            break;
-
-        default:
-            break;
-    }
-
-    /* Print extended information if appropriate */
-    print_attr_msg(data->stream, data->node, data->node->details->running_rsc,
-                   name, value);
-
-    /* Close out the attribute */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("\n");
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(data->stream, "</li>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(data->stream, " />\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-static void
-print_node_summary(FILE *stream, pe_working_set_t * data_set, gboolean operations)
-{
-    xmlNode *node_state = NULL;
-    xmlNode *cib_status = get_object_root(XML_CIB_TAG_STATUS, data_set->input);
-
-    /* Print heading */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            if (operations) {
-                print_as("\nOperations:\n");
-            } else {
-                print_as("\nMigration Summary:\n");
-            }
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            if (operations) {
-                fprintf(stream, " <hr />\n <h2>Operations</h2>\n");
-            } else {
-                fprintf(stream, " <hr />\n <h2>Migration Summary</h2>\n");
-            }
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    <node_history>\n");
-            break;
-
-        default:
-            break;
-    }
-
-    /* Print each node in the CIB status */
-    for (node_state = __xml_first_child(cib_status); node_state != NULL;
-         node_state = __xml_next(node_state)) {
-        if (crm_str_eq((const char *)node_state->name, XML_CIB_TAG_STATE, TRUE)) {
-            print_node_history(stream, data_set, node_state, operations);
-        }
-    }
-
-    /* Close section */
-    switch (output_format) {
-        case mon_output_xml:
-            fprintf(stream, "    </node_history>\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-static void
-print_ticket(gpointer name, gpointer value, gpointer data)
-{
-    ticket_t *ticket = (ticket_t *) value;
-    FILE *stream = (FILE *) data;
-
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("* %s:\t%s%s", ticket->id,
-                     (ticket->granted? "granted" : "revoked"),
-                     (ticket->standby? " [standby]" : ""));
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, "  <li>%s: %s%s", ticket->id,
-                    (ticket->granted? "granted" : "revoked"),
-                    (ticket->standby? " [standby]" : ""));
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "        <ticket id=\"%s\" status=\"%s\" standby=\"%s\"",
-                    ticket->id, (ticket->granted? "granted" : "revoked"),
-                    (ticket->standby? "true" : "false"));
-            break;
-
-        default:
-            break;
-    }
-    if (ticket->last_granted > -1) {
-        print_nvpair(stdout, "last-granted", NULL, NULL, ticket->last_granted);
-    }
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("\n");
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, "</li>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, " />\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-static void
-print_cluster_tickets(FILE *stream, pe_working_set_t * data_set)
-{
-    /* Print section heading */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("\nTickets:\n");
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " <hr />\n <h2>Tickets</h2>\n <ul>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    <tickets>\n");
-            break;
-
-        default:
-            break;
-    }
-
-    /* Print each ticket */
-    g_hash_table_foreach(data_set->tickets, print_ticket, stream);
-
-    /* Close section */
-    switch (output_format) {
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " </ul>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    </tickets>\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Return human-friendly string representing node name
- *
- * The returned string will be in the format
- *    uname[@hostUname] [(nodeID)]
- * "@hostUname" will be printed if the node is a guest node.
- * "(nodeID)" will be printed if the node ID is different from the node uname,
- *  and detailed output has been requested.
- *
- * \param[in] node  Node to represent
- * \return Newly allocated string with representation of node name
- * \note It is the caller's responsibility to free the result with free().
- */
-static char *
-get_node_display_name(node_t *node)
-{
-    char *node_name;
-    const char *node_host = NULL;
-    const char *node_id = NULL;
-    int name_len;
-
-    CRM_ASSERT((node != NULL) && (node->details != NULL) && (node->details->uname != NULL));
-
-    /* Host is displayed only if this is a guest node */
-    if (is_container_remote_node(node)) {
-        node_t *host_node = pe__current_node(node->details->remote_rsc);
-
-        if (host_node && host_node->details) {
-            node_host = host_node->details->uname;
-        }
-        if (node_host == NULL) {
-            node_host = ""; /* so we at least get "uname@" to indicate guest */
-        }
-    }
-
-    /* Node ID is displayed if different from uname and detail is requested */
-    if (print_clone_detail && safe_str_neq(node->details->uname, node->details->id)) {
-        node_id = node->details->id;
-    }
-
-    /* Determine name length */
-    name_len = strlen(node->details->uname) + 1;
-    if (node_host) {
-        name_len += strlen(node_host) + 1; /* "@node_host" */
-    }
-    if (node_id) {
-        name_len += strlen(node_id) + 3; /* + " (node_id)" */
-    }
-
-    /* Allocate and populate display name */
-    node_name = malloc(name_len);
-    CRM_ASSERT(node_name != NULL);
-    strcpy(node_name, node->details->uname);
-    if (node_host) {
-        strcat(node_name, "@");
-        strcat(node_name, node_host);
-    }
-    if (node_id) {
-        strcat(node_name, " (");
-        strcat(node_name, node_id);
-        strcat(node_name, ")");
-    }
-    return node_name;
-}
-
-/*!
- * \internal
- * \brief Print a negative location constraint
- *
- * \param[in] stream     File stream to display output to
- * \param[in] node       Node affected by constraint
- * \param[in] location   Constraint to print
- */
-static void
-print_ban(FILE *stream, pe_node_t *node, pe__location_t *location)
-{
-    char *node_name = NULL;
-
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            node_name = get_node_display_name(node);
-            print_as(" %s\tprevents %s from running %son %s\n",
-                     location->id, location->rsc_lh->id,
-                     ((location->role_filter == RSC_ROLE_MASTER)? "as Master " : ""),
-                     node_name);
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            node_name = get_node_display_name(node);
-            fprintf(stream, "  <li>%s prevents %s from running %son %s</li>\n",
-                     location->id, location->rsc_lh->id,
-                     ((location->role_filter == RSC_ROLE_MASTER)? "as Master " : ""),
-                     node_name);
-            break;
-
-        case mon_output_xml:
-            fprintf(stream,
-                    "        <ban id=\"%s\" resource=\"%s\" node=\"%s\" weight=\"%d\" master_only=\"%s\" />\n",
-                    location->id, location->rsc_lh->id, node->details->uname, node->weight,
-                    ((location->role_filter == RSC_ROLE_MASTER)? "true" : "false"));
-            break;
-
-        default:
-            break;
-    }
-    free(node_name);
-}
-
-/*!
- * \internal
- * \brief Print section for negative location constraints
- *
- * \param[in] stream     File stream to display output to
- * \param[in] data_set   Working set corresponding to CIB status to display
- */
-static void
-print_neg_locations(FILE *stream, pe_working_set_t *data_set)
-{
-    GListPtr gIter, gIter2;
-
-    /* Print section heading */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("\nNegative Location Constraints:\n");
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " <hr />\n <h2>Negative Location Constraints</h2>\n <ul>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    <bans>\n");
-            break;
-
-        default:
-            break;
-    }
-
-    /* Print each ban */
-    for (gIter = data_set->placement_constraints; gIter != NULL; gIter = gIter->next) {
-        pe__location_t *location = gIter->data;
-        if (!g_str_has_prefix(location->id, print_neg_location_prefix))
-            continue;
-        for (gIter2 = location->node_list_rh; gIter2 != NULL; gIter2 = gIter2->next) {
-            node_t *node = (node_t *) gIter2->data;
-
-            if (node->weight < 0) {
-                print_ban(stream, node, location);
-            }
-        }
-    }
-
-    /* Close section */
-    switch (output_format) {
-        case mon_output_cgi:
-        case mon_output_html:
-            fprintf(stream, " </ul>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    </bans>\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-static void
-crm_mon_get_parameters(resource_t *rsc, pe_working_set_t * data_set)
-{
-    get_rsc_attributes(rsc->parameters, rsc, NULL, data_set);
-    crm_trace("Beekhof: unpacked params for %s (%d)", rsc->id, g_hash_table_size(rsc->parameters));
-    if(rsc->children) {
-        GListPtr gIter = NULL;
-
-        for (gIter = rsc->children; gIter != NULL; gIter = gIter->next) {
-            crm_mon_get_parameters(gIter->data, data_set);
-        }
-    }
-}
-
-/*!
- * \internal
- * \brief Print node attributes section
- *
- * \param[in] stream     File stream to display output to
- * \param[in] data_set   Working set of CIB state
- */
-static void
-print_node_attributes(FILE *stream, pe_working_set_t *data_set)
-{
-    GListPtr gIter = NULL;
-
-    /* Print section heading */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("\nNode Attributes:\n");
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " <hr />\n <h2>Node Attributes</h2>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    <node_attributes>\n");
-            break;
-
-        default:
-            break;
-    }
-
-    /* Unpack all resource parameters (it would be more efficient to do this
-     * only when needed for the first time in print_attr_msg())
-     */
-    for (gIter = data_set->resources; gIter != NULL; gIter = gIter->next) {
-        crm_mon_get_parameters(gIter->data, data_set);
-    }
-
-    /* Display each node's attributes */
-    for (gIter = data_set->nodes; gIter != NULL; gIter = gIter->next) {
-        struct mon_attr_data data;
-
-        data.stream = stream;
-        data.node = (node_t *) gIter->data;
-
-        if (data.node && data.node->details && data.node->details->online) {
-            print_node_start(stream, data.node);
-            g_hash_table_foreach(data.node->details->attrs, create_attr_list, NULL);
-            g_list_foreach(attr_list, print_node_attribute, &data);
-            g_list_free(attr_list);
-            attr_list = NULL;
-            print_node_end(stream);
-        }
-    }
-
-    /* Print section footer */
-    switch (output_format) {
-        case mon_output_xml:
-            fprintf(stream, "    </node_attributes>\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Return resource display options corresponding to command-line choices
- *
- * \return Bitmask of pe_print_options suitable for resource print functions
- */
-static int
-get_resource_display_options(void)
-{
-    int print_opts;
-
-    /* Determine basic output format */
-    switch (output_format) {
-        case mon_output_console:
-            print_opts = pe_print_ncurses;
-            break;
-        case mon_output_html:
-        case mon_output_cgi:
-            print_opts = pe_print_html;
-            break;
-        case mon_output_xml:
-            print_opts = pe_print_xml;
-            break;
-        default:
-            print_opts = pe_print_printf;
-            break;
-    }
-
-    /* Add optional display elements */
-    if (print_pending) {
-        print_opts |= pe_print_pending;
-    }
-    if (print_clone_detail) {
-        print_opts |= pe_print_clone_details|pe_print_implicit;
-    }
-    if (!inactive_resources) {
-        print_opts |= pe_print_clone_active;
-    }
-    if (print_brief) {
-        print_opts |= pe_print_brief;
-    }
-    return print_opts;
-}
-
-/*!
- * \internal
- * \brief Return human-friendly string representing current time
- *
- * \return Current time as string (as by ctime() but without newline) on success
- *         or "Could not determine current time" on error
- * \note The return value points to a statically allocated string which might be
- *       overwritten by subsequent calls to any of the C library date and time functions.
- */
-static const char *
-crm_now_string(void)
-{
-    time_t a_time = time(NULL);
-    char *since_epoch = ctime(&a_time);
-
-    if ((a_time == (time_t) -1) || (since_epoch == NULL)) {
-        return "Could not determine current time";
-    }
-    since_epoch[strlen(since_epoch) - 1] = EOS; /* trim newline */
-    return (since_epoch);
-}
-
-/*!
- * \internal
- * \brief Print header for cluster summary if needed
- *
- * \param[in] stream     File stream to display output to
- */
-static void
-print_cluster_summary_header(FILE *stream)
-{
-    switch (output_format) {
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " <h2>Cluster Summary</h2>\n <p>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    <summary>\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print footer for cluster summary if needed
- *
- * \param[in] stream     File stream to display output to
- */
-static void
-print_cluster_summary_footer(FILE *stream)
-{
-    switch (output_format) {
-        case mon_output_cgi:
-        case mon_output_html:
-            fprintf(stream, " </p>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    </summary>\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print times the display was last updated and CIB last changed
- *
- * \param[in] stream     File stream to display output to
- * \param[in] data_set   Working set of CIB state
- */
-static void
-print_cluster_times(FILE *stream, pe_working_set_t *data_set)
-{
-    const char *last_written = crm_element_value(data_set->input, XML_CIB_ATTR_WRITTEN);
-    const char *user = crm_element_value(data_set->input, XML_ATTR_UPDATE_USER);
-    const char *client = crm_element_value(data_set->input, XML_ATTR_UPDATE_CLIENT);
-    const char *origin = crm_element_value(data_set->input, XML_ATTR_UPDATE_ORIG);
-
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("Last updated: %s", crm_now_string());
-            print_as((user || client || origin)? "\n" : "\t\t");
-            print_as("Last change: %s", last_written ? last_written : "");
-            if (user) {
-                print_as(" by %s", user);
-            }
-            if (client) {
-                print_as(" via %s", client);
-            }
-            if (origin) {
-                print_as(" on %s", origin);
-            }
-            print_as("\n");
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " <b>Last updated:</b> %s<br/>\n", crm_now_string());
-            fprintf(stream, " <b>Last change:</b> %s", last_written ? last_written : "");
-            if (user) {
-                fprintf(stream, " by %s", user);
-            }
-            if (client) {
-                fprintf(stream, " via %s", client);
-            }
-            if (origin) {
-                fprintf(stream, " on %s", origin);
-            }
-            fprintf(stream, "<br/>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "        <last_update time=\"%s\" />\n", crm_now_string());
-            fprintf(stream, "        <last_change time=\"%s\" user=\"%s\" client=\"%s\" origin=\"%s\" />\n",
-                    last_written ? last_written : "", user ? user : "",
-                    client ? client : "", origin ? origin : "");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print cluster stack
- *
- * \param[in] stream     File stream to display output to
- * \param[in] stack_s    Stack name
- */
-static void
-print_cluster_stack(FILE *stream, const char *stack_s)
-{
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("Stack: %s\n", stack_s);
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " <b>Stack:</b> %s<br/>\n", stack_s);
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "        <stack type=\"%s\" />\n", stack_s);
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print current DC and its version
- *
- * \param[in] stream     File stream to display output to
- * \param[in] data_set   Working set of CIB state
- */
-static void
-print_cluster_dc(FILE *stream, pe_working_set_t *data_set)
-{
-    node_t *dc = data_set->dc_node;
-    xmlNode *dc_version = get_xpath_object("//nvpair[@name='dc-version']",
-                                           data_set->input, LOG_DEBUG);
-    const char *dc_version_s = dc_version?
-                               crm_element_value(dc_version, XML_NVPAIR_ATTR_VALUE)
-                               : NULL;
-    const char *quorum = crm_element_value(data_set->input, XML_ATTR_HAVE_QUORUM);
-    char *dc_name = dc? get_node_display_name(dc) : NULL;
-
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("Current DC: ");
-            if (dc) {
-                print_as("%s (version %s) - partition %s quorum\n",
-                         dc_name, (dc_version_s? dc_version_s : "unknown"),
-                         (crm_is_true(quorum) ? "with" : "WITHOUT"));
-            } else {
-                print_as("NONE\n");
-            }
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " <b>Current DC:</b> ");
-            if (dc) {
-                fprintf(stream, "%s (version %s) - partition %s quorum",
-                        dc_name, (dc_version_s? dc_version_s : "unknown"),
-                        (crm_is_true(quorum)? "with" : "<font color=\"red\"><b>WITHOUT</b></font>"));
-            } else {
-                fprintf(stream, "<font color=\"red\"><b>NONE</b></font>");
-            }
-            fprintf(stream, "<br/>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream,  "        <current_dc ");
-            if (dc) {
-                fprintf(stream,
-                        "present=\"true\" version=\"%s\" name=\"%s\" id=\"%s\" with_quorum=\"%s\"",
-                        (dc_version_s? dc_version_s : ""), dc->details->uname, dc->details->id,
-                        (crm_is_true(quorum) ? "true" : "false"));
-            } else {
-                fprintf(stream, "present=\"false\"");
-            }
-            fprintf(stream, " />\n");
-            break;
-
-        default:
-            break;
-    }
-    free(dc_name);
-}
-
-/*!
- * \internal
- * \brief Print counts of configured nodes and resources
- *
- * \param[in] stream     File stream to display output to
- * \param[in] data_set   Working set of CIB state
- * \param[in] stack_s    Stack name
- */
-static void
-print_cluster_counts(FILE *stream, pe_working_set_t *data_set, const char *stack_s)
-{
-    int nnodes = g_list_length(data_set->nodes);
-    int nresources = count_resources(data_set, NULL);
-
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-
-            print_as("\n%d node%s configured\n", nnodes, s_if_plural(nnodes));
-
-            print_as("%d resource%s configured",
-                     nresources, s_if_plural(nresources));
-            if(data_set->disabled_resources || data_set->blocked_resources) {
-                print_as(" (");
-                if (data_set->disabled_resources) {
-                    print_as("%d DISABLED", data_set->disabled_resources);
-                }
-                if (data_set->disabled_resources && data_set->blocked_resources) {
-                    print_as(", ");
-                }
-                if (data_set->blocked_resources) {
-                    print_as("%d BLOCKED from starting due to failure",
-                             data_set->blocked_resources);
-                }
-                print_as(")");
-            }
-            print_as("\n");
-
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-
-            fprintf(stream, " %d node%s configured<br/>\n",
-                    nnodes, s_if_plural(nnodes));
-
-            fprintf(stream, " %d resource%s configured",
-                    nresources, s_if_plural(nresources));
-            if (data_set->disabled_resources || data_set->blocked_resources) {
-                fprintf(stream, " (");
-                if (data_set->disabled_resources) {
-                    fprintf(stream, "%d <strong>DISABLED</strong>",
-                            data_set->disabled_resources);
-                }
-                if (data_set->disabled_resources && data_set->blocked_resources) {
-                    fprintf(stream, ", ");
-                }
-                if (data_set->blocked_resources) {
-                    fprintf(stream,
-                            "%d <strong>BLOCKED</strong> from starting due to failure",
-                            data_set->blocked_resources);
-                }
-                fprintf(stream, ")");
-            }
-            fprintf(stream, "<br/>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream,
-                    "        <nodes_configured number=\"%d\" />\n",
-                    g_list_length(data_set->nodes));
-            fprintf(stream,
-                    "        <resources_configured number=\"%d\" disabled=\"%d\" blocked=\"%d\" />\n",
-                    count_resources(data_set, NULL),
-                    data_set->disabled_resources, data_set->blocked_resources);
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print cluster-wide options
- *
- * \param[in] stream     File stream to display output to
- * \param[in] data_set   Working set of CIB state
- *
- * \note Currently this is only implemented for HTML and XML output, and
- *       prints only a few options. If there is demand, more could be added.
- */
-static void
-print_cluster_options(FILE *stream, pe_working_set_t *data_set)
-{
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            if (is_set(data_set->flags, pe_flag_maintenance_mode)) {
-                print_as("\n              *** Resource management is DISABLED ***");
-                print_as("\n  The cluster will not attempt to start, stop or recover services");
-                print_as("\n");
-            }
-            break;
-
-        case mon_output_html:
-            fprintf(stream, " </p>\n <h3>Config Options</h3>\n");
-            fprintf(stream, " <table>\n");
-            fprintf(stream, "  <tr><th>STONITH of failed nodes</th><td>%s</td></tr>\n",
-                    is_set(data_set->flags, pe_flag_stonith_enabled)? "enabled" : "disabled");
-
-            fprintf(stream, "  <tr><th>Cluster is</th><td>%ssymmetric</td></tr>\n",
-                    is_set(data_set->flags, pe_flag_symmetric_cluster)? "" : "a");
-
-            fprintf(stream, "  <tr><th>No Quorum Policy</th><td>");
-            switch (data_set->no_quorum_policy) {
-                case no_quorum_freeze:
-                    fprintf(stream, "Freeze resources");
-                    break;
-                case no_quorum_stop:
-                    fprintf(stream, "Stop ALL resources");
-                    break;
-                case no_quorum_ignore:
-                    fprintf(stream, "Ignore");
-                    break;
-                case no_quorum_suicide:
-                    fprintf(stream, "Suicide");
-                    break;
-            }
-            fprintf(stream, "</td></tr>\n");
-
-            fprintf(stream, "  <tr><th>Resource management</th><td>");
-            if (is_set(data_set->flags, pe_flag_maintenance_mode)) {
-                fprintf(stream, "<strong>DISABLED</strong> (the cluster will "
-                                "not attempt to start, stop or recover services)");
-            } else {
-                fprintf(stream, "enabled");
-            }
-            fprintf(stream, "</td></tr>\n");
-
-            fprintf(stream, "</table>\n <p>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "        <cluster_options");
-            fprintf(stream, " stonith-enabled=\"%s\"",
-                    is_set(data_set->flags, pe_flag_stonith_enabled)?
-                    "true" : "false");
-            fprintf(stream, " symmetric-cluster=\"%s\"",
-                    is_set(data_set->flags, pe_flag_symmetric_cluster)?
-                    "true" : "false");
-            fprintf(stream, " no-quorum-policy=\"");
-            switch (data_set->no_quorum_policy) {
-                case no_quorum_freeze:
-                    fprintf(stream, "freeze");
-                    break;
-                case no_quorum_stop:
-                    fprintf(stream, "stop");
-                    break;
-                case no_quorum_ignore:
-                    fprintf(stream, "ignore");
-                    break;
-                case no_quorum_suicide:
-                    fprintf(stream, "suicide");
-                    break;
-            }
-            fprintf(stream, "\"");
-            fprintf(stream, " maintenance-mode=\"%s\"",
-                    is_set(data_set->flags, pe_flag_maintenance_mode)?
-                    "true" : "false");
-            fprintf(stream, " />\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Get the name of the stack in use (or "unknown" if not available)
- *
- * \param[in] data_set   Working set of CIB state
- *
- * \return String representing stack name
- */
-static const char *
-get_cluster_stack(pe_working_set_t *data_set)
-{
-    xmlNode *stack = get_xpath_object("//nvpair[@name='cluster-infrastructure']",
-                                      data_set->input, LOG_DEBUG);
-    return stack? crm_element_value(stack, XML_NVPAIR_ATTR_VALUE) : "unknown";
-}
-
-/*!
- * \internal
- * \brief Print a summary of cluster-wide information
- *
- * \param[in] stream     File stream to display output to
- * \param[in] data_set   Working set of CIB state
- */
-static void
-print_cluster_summary(FILE *stream, pe_working_set_t *data_set)
-{
-    const char *stack_s = get_cluster_stack(data_set);
-    gboolean header_printed = FALSE;
-
-    if (show & mon_show_stack) {
-        if (header_printed == FALSE) {
-            print_cluster_summary_header(stream);
-            header_printed = TRUE;
-        }
-        print_cluster_stack(stream, stack_s);
-    }
-
-    /* Always print DC if none, even if not requested */
-    if ((data_set->dc_node == NULL) || (show & mon_show_dc)) {
-        if (header_printed == FALSE) {
-            print_cluster_summary_header(stream);
-            header_printed = TRUE;
-        }
-        print_cluster_dc(stream, data_set);
-    }
-
-    if (show & mon_show_times) {
-        if (header_printed == FALSE) {
-            print_cluster_summary_header(stream);
-            header_printed = TRUE;
-        }
-        print_cluster_times(stream, data_set);
-    }
-
-    if (is_set(data_set->flags, pe_flag_maintenance_mode)
-        || data_set->disabled_resources
-        || data_set->blocked_resources
-        || is_set(show, mon_show_count)) {
-        if (header_printed == FALSE) {
-            print_cluster_summary_header(stream);
-            header_printed = TRUE;
-        }
-        print_cluster_counts(stream, data_set, stack_s);
-    }
-
-    /* There is not a separate option for showing cluster options, so show with
-     * stack for now; a separate option could be added if there is demand
-     */
-    if (show & mon_show_stack) {
-        print_cluster_options(stream, data_set);
-    }
-
-    if (header_printed) {
-        print_cluster_summary_footer(stream);
-    }
-}
-
-/*!
- * \internal
- * \brief Print a failed action
- *
- * \param[in] stream     File stream to display output to
- * \param[in] xml_op     Root of XML tree describing failed action
- */
-static void
-print_failed_action(FILE *stream, xmlNode *xml_op)
-{
-    const char *op_key = crm_element_value(xml_op, XML_LRM_ATTR_TASK_KEY);
-    const char *op_key_attr = "op_key";
-    const char *last = crm_element_value(xml_op, XML_RSC_OP_LAST_CHANGE);
-    const char *node = crm_element_value(xml_op, XML_ATTR_UNAME);
-    const char *call = crm_element_value(xml_op, XML_LRM_ATTR_CALLID);
-    const char *exit_reason = crm_element_value(xml_op, XML_LRM_ATTR_EXIT_REASON);
-    int rc = crm_parse_int(crm_element_value(xml_op, XML_LRM_ATTR_RC), "0");
-    int status = crm_parse_int(crm_element_value(xml_op, XML_LRM_ATTR_OPSTATUS), "0");
-    char *exit_reason_cleaned;
-
-    /* If no op_key was given, use id instead */
-    if (op_key == NULL) {
-        op_key = ID(xml_op);
-        op_key_attr = "id";
-    }
-
-    /* If no exit reason was given, use "none" */
-    if (exit_reason == NULL) {
-        exit_reason = "none";
-    }
-
-    /* Print common action information */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("* %s on %s '%s' (%d): call=%s, status=%s, exitreason='%s'",
-                     op_key, node, services_ocf_exitcode_str(rc), rc,
-                     call, services_lrm_status_str(status), exit_reason);
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, "  <li>%s on %s '%s' (%d): call=%s, status=%s, exitreason='%s'",
-                     op_key, node, services_ocf_exitcode_str(rc), rc,
-                     call, services_lrm_status_str(status), exit_reason);
-            break;
-
-        case mon_output_xml:
-            exit_reason_cleaned = crm_xml_escape(exit_reason);
-            fprintf(stream, "        <failure %s=\"%s\" node=\"%s\"",
-                    op_key_attr, op_key, node);
-            fprintf(stream, " exitstatus=\"%s\" exitreason=\"%s\" exitcode=\"%d\"",
-                    services_ocf_exitcode_str(rc), exit_reason_cleaned, rc);
-            fprintf(stream, " call=\"%s\" status=\"%s\"",
-                    call, services_lrm_status_str(status));
-            free(exit_reason_cleaned);
-            break;
-
-        default:
-            break;
-    }
-
-    /* If last change was given, print timing information as well */
-    if (last) {
-        time_t run_at = crm_parse_int(last, "0");
-        char *run_at_s = ctime(&run_at);
-
-        if (run_at_s) {
-            run_at_s[24] = 0; /* Overwrite the newline */
-        }
-
-        switch (output_format) {
-            case mon_output_plain:
-            case mon_output_console:
-                print_as(",\n    last-rc-change='%s', queued=%sms, exec=%sms",
-                         run_at_s? run_at_s : "",
-                         crm_element_value(xml_op, XML_RSC_OP_T_QUEUE),
-                         crm_element_value(xml_op, XML_RSC_OP_T_EXEC));
-                break;
-
-            case mon_output_html:
-            case mon_output_cgi:
-                fprintf(stream, " last-rc-change='%s', queued=%sms, exec=%sms",
-                        run_at_s? run_at_s : "",
-                        crm_element_value(xml_op, XML_RSC_OP_T_QUEUE),
-                        crm_element_value(xml_op, XML_RSC_OP_T_EXEC));
-                break;
-
-            case mon_output_xml:
-                fprintf(stream,
-                        " last-rc-change=\"%s\" queued=\"%s\" exec=\"%s\" interval=\"%u\" task=\"%s\"",
-                        run_at_s? run_at_s : "",
-                        crm_element_value(xml_op, XML_RSC_OP_T_QUEUE),
-                        crm_element_value(xml_op, XML_RSC_OP_T_EXEC),
-                        crm_parse_ms(crm_element_value(xml_op, XML_LRM_ATTR_INTERVAL_MS)),
-                        crm_element_value(xml_op, XML_LRM_ATTR_TASK));
-                break;
-
-            default:
-                break;
-        }
-    }
-
-    /* End the action listing */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("\n");
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, "</li>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, " />\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
- * \brief Print a section for failed actions
- *
- * \param[in] stream     File stream to display output to
- * \param[in] data_set   Working set of CIB state
- */
-static void
-print_failed_actions(FILE *stream, pe_working_set_t *data_set)
-{
-    xmlNode *xml_op = NULL;
-
-    /* Print section heading */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("\nFailed Resource Actions:\n");
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream,
-                    " <hr />\n <h2>Failed Resource Actions</h2>\n <ul>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    <failures>\n");
-            break;
-
-        default:
-            break;
-    }
-
-    /* Print each failed action */
-    for (xml_op = __xml_first_child(data_set->failed); xml_op != NULL;
-         xml_op = __xml_next(xml_op)) {
-        print_failed_action(stream, xml_op);
-    }
-
-    /* End section */
-    switch (output_format) {
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " </ul>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    </failures>\n");
-            break;
-
-        default:
-            break;
-    }
-}
 
 /*!
  * \internal
@@ -3114,116 +1053,6 @@ sort_stonith_history(stonith_history_t *history)
 
 /*!
  * \internal
- * \brief Print a stonith action
- *
- * \param[in] stream     File stream to display output to
- * \param[in] event      stonith event
- */
-static void
-print_stonith_action(FILE *stream, stonith_history_t *event)
-{
-    const char *action_s = stonith_action_str(event->action);
-    char *run_at_s = ctime(&event->completed);
-
-    if ((run_at_s) && (run_at_s[0] != 0)) {
-        run_at_s[strlen(run_at_s)-1] = 0; /* Overwrite the newline */
-    }
-
-    switch(output_format) {
-        case mon_output_xml:
-            fprintf(stream, "        <fence_event target=\"%s\" action=\"%s\"",
-                    event->target, event->action);
-            switch(event->state) {
-                case st_done:
-                    fprintf(stream, " state=\"success\"");
-                    break;
-                case st_failed:
-                    fprintf(stream, " state=\"failed\"");
-                    break;
-                default:
-                    fprintf(stream, " state=\"pending\"");
-            }
-            fprintf(stream, " origin=\"%s\" client=\"%s\"",
-                    event->origin, event->client);
-            if (event->delegate) {
-                fprintf(stream, " delegate=\"%s\"", event->delegate);
-            }
-            switch(event->state) {
-                case st_done:
-                case st_failed:
-                    fprintf(stream, " completed=\"%s\"", run_at_s?run_at_s:"");
-                    break;
-                default:
-                    break;
-            }
-            fprintf(stream, " />\n");
-            break;
-
-        case mon_output_plain:
-        case mon_output_console:
-            switch(event->state) {
-                case st_done:
-                    print_as("* %s of %s successful: delegate=%s, client=%s, origin=%s,\n"
-                             "    %s='%s'\n",
-                             action_s, event->target,
-                             event->delegate ? event->delegate : "",
-                             event->client, event->origin,
-                             fence_full_history?"completed":"last-successful",
-                             run_at_s?run_at_s:"");
-                    break;
-                case st_failed:
-                    print_as("* %s of %s failed: delegate=%s, client=%s, origin=%s,\n"
-                             "    %s='%s'\n",
-                             action_s, event->target,
-                             event->delegate ? event->delegate : "",
-                             event->client, event->origin,
-                             fence_full_history?"completed":"last-failed",
-                             run_at_s?run_at_s:"");
-                    break;
-                default:
-                    print_as("* %s of %s pending: client=%s, origin=%s\n",
-                             action_s, event->target,
-                             event->client, event->origin);
-            }
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            switch(event->state) {
-                case st_done:
-                    fprintf(stream, "  <li>%s of %s successful: delegate=%s, "
-                                    "client=%s, origin=%s, %s='%s'</li>\n",
-                                    action_s, event->target,
-                                    event->delegate ? event->delegate : "",
-                                    event->client, event->origin,
-                                    fence_full_history?"completed":"last-successful",
-                                    run_at_s?run_at_s:"");
-                    break;
-                case st_failed:
-                    fprintf(stream, "  <li>%s of %s failed: delegate=%s, "
-                                    "client=%s, origin=%s, %s='%s'</li>\n",
-                                    action_s, event->target,
-                                    event->delegate ? event->delegate : "",
-                                    event->client, event->origin,
-                                    fence_full_history?"completed":"last-failed",
-                                    run_at_s?run_at_s:"");
-                    break;
-                default:
-                    fprintf(stream, "  <li>%s of %s pending: client=%s, "
-                                    "origin=%s</li>\n",
-                                    action_s, event->target,
-                                    event->client, event->origin);
-            }
-            break;
-
-        default:
-            /* no support for fence history for other formats so far */
-            break;
-    }
-}
-
-/*!
- * \internal
  * \brief Print a section for failed stonith actions
  *
  * \param[in] stream     File stream to display output to
@@ -3339,61 +1168,6 @@ print_stonith_pending(FILE *stream, stonith_history_t *history)
 
 /*!
  * \internal
- * \brief Print a section for stonith-history
- *
- * \param[in] stream     File stream to display output to
- * \param[in] history    List of stonith actions
- *
- */
-static void
-print_stonith_history(FILE *stream, stonith_history_t *history)
-{
-    stonith_history_t *hp;
-
-    /* Print section heading */
-    switch (output_format) {
-        case mon_output_plain:
-        case mon_output_console:
-            print_as("\nFencing History:\n");
-            break;
-
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " <hr />\n <h2>Fencing History</h2>\n <ul>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    <fence_history>\n");
-            break;
-
-        default:
-            break;
-    }
-
-    for (hp = history; hp; hp = hp->next) {
-        if ((hp->state != st_failed) || (output_format == mon_output_xml)) {
-            print_stonith_action(stream, hp);
-        }
-    }
-
-    /* End section */
-    switch (output_format) {
-        case mon_output_html:
-        case mon_output_cgi:
-            fprintf(stream, " </ul>\n");
-            break;
-
-        case mon_output_xml:
-            fprintf(stream, "    </fence_history>\n");
-            break;
-
-        default:
-            break;
-    }
-}
-
-/*!
- * \internal
  * \brief Print cluster status to screen
  *
  * This uses the global display preferences set by command-line options
@@ -3466,7 +1240,7 @@ print_status(pe_working_set_t * data_set,
 
         } else if (node->details->online) {
             node_mode = "online";
-            if (group_by_node == FALSE) {
+            if (pp.group_by_node == FALSE) {
                 if (is_container_remote_node(node)) {
                     online_guest_nodes = add_list_element(online_guest_nodes, node_name);
                 } else if (is_baremetal_remote_node(node)) {
@@ -3479,7 +1253,7 @@ print_status(pe_working_set_t * data_set,
             }
         } else {
             node_mode = "OFFLINE";
-            if (group_by_node == FALSE) {
+            if (pp.group_by_node == FALSE) {
                 if (is_baremetal_remote_node(node)) {
                     offline_remote_nodes = add_list_element(offline_remote_nodes, node_name);
                 } else if (is_container_remote_node(node)) {
@@ -3503,8 +1277,8 @@ print_status(pe_working_set_t * data_set,
         print_as("Node %s: %s\n", node_name, node_mode);
 
         /* If we're grouping by node, print its resources */
-        if (group_by_node) {
-            if (print_brief) {
+        if (pp.group_by_node) {
+            if (pp.print_brief) {
                 print_rscs_brief(node->details->running_rsc, "\t", print_opts | pe_print_rsconly,
                                  stdout, FALSE);
             } else {
@@ -3546,16 +1320,16 @@ print_status(pe_working_set_t * data_set,
     print_resources(stdout, data_set, print_opts);
 
     /* print Node Attributes section if requested */
-    if (show & mon_show_attributes) {
+    if (pp.show & mon_show_attributes) {
         print_node_attributes(stdout, data_set);
     }
 
     /* If requested, print resource operations (which includes failcounts)
      * or just failcounts
      */
-    if (show & (mon_show_operations | mon_show_failcounts)) {
+    if (pp.show & (mon_show_operations | mon_show_failcounts)) {
         print_node_summary(stdout, data_set,
-                           ((show & mon_show_operations)? TRUE : FALSE));
+                           ((pp.show & mon_show_operations)? TRUE : FALSE));
     }
 
     /* If there were any failed actions, print them */
@@ -3564,23 +1338,23 @@ print_status(pe_working_set_t * data_set,
     }
 
     /* Print failed stonith actions */
-    if (fence_history) {
+    if (pp.fence_history) {
         print_failed_stonith_actions(stdout, stonith_history);
     }
 
     /* Print tickets if requested */
-    if (show & mon_show_tickets) {
+    if (pp.show & mon_show_tickets) {
         print_cluster_tickets(stdout, data_set);
     }
 
     /* Print negative location constraints if requested */
-    if (show & mon_show_bans) {
+    if (pp.show & mon_show_bans) {
         print_neg_locations(stdout, data_set);
     }
 
     /* Print stonith history */
-    if (fence_history) {
-        if (show & mon_show_fence_history) {
+    if (pp.fence_history) {
+        if (pp.show & mon_show_fence_history) {
             print_stonith_history(stdout, stonith_history);
         } else {
             print_stonith_pending(stdout, stonith_history);
@@ -3592,117 +1366,6 @@ print_status(pe_working_set_t * data_set,
         refresh();
     }
 #endif
-}
-
-/*!
- * \internal
- * \brief Print cluster status in XML format
- *
- * \param[in] data_set   Working set of CIB state
- */
-static void
-print_xml_status(pe_working_set_t * data_set,
-                 stonith_history_t *stonith_history)
-{
-    FILE *stream = stdout;
-    GListPtr gIter = NULL;
-    int print_opts = get_resource_display_options();
-
-    fprintf(stream, "<?xml version=\"1.0\"?>\n");
-    fprintf(stream, "<crm_mon version=\"%s\">\n", VERSION);
-
-    print_cluster_summary(stream, data_set);
-
-    /*** NODES ***/
-    fprintf(stream, "    <nodes>\n");
-    for (gIter = data_set->nodes; gIter != NULL; gIter = gIter->next) {
-        node_t *node = (node_t *) gIter->data;
-        const char *node_type = "unknown";
-
-        switch (node->details->type) {
-            case node_member:
-                node_type = "member";
-                break;
-            case node_remote:
-                node_type = "remote";
-                break;
-            case node_ping:
-                node_type = "ping";
-                break;
-        }
-
-        fprintf(stream, "        <node name=\"%s\" ", node->details->uname);
-        fprintf(stream, "id=\"%s\" ", node->details->id);
-        fprintf(stream, "online=\"%s\" ", node->details->online ? "true" : "false");
-        fprintf(stream, "standby=\"%s\" ", node->details->standby ? "true" : "false");
-        fprintf(stream, "standby_onfail=\"%s\" ", node->details->standby_onfail ? "true" : "false");
-        fprintf(stream, "maintenance=\"%s\" ", node->details->maintenance ? "true" : "false");
-        fprintf(stream, "pending=\"%s\" ", node->details->pending ? "true" : "false");
-        fprintf(stream, "unclean=\"%s\" ", node->details->unclean ? "true" : "false");
-        fprintf(stream, "shutdown=\"%s\" ", node->details->shutdown ? "true" : "false");
-        fprintf(stream, "expected_up=\"%s\" ", node->details->expected_up ? "true" : "false");
-        fprintf(stream, "is_dc=\"%s\" ", node->details->is_dc ? "true" : "false");
-        fprintf(stream, "resources_running=\"%d\" ", g_list_length(node->details->running_rsc));
-        fprintf(stream, "type=\"%s\" ", node_type);
-        if (is_container_remote_node(node)) {
-            fprintf(stream, "id_as_resource=\"%s\" ", node->details->remote_rsc->container->id);
-        }
-
-        if (group_by_node) {
-            GListPtr lpc2 = NULL;
-
-            fprintf(stream, ">\n");
-            for (lpc2 = node->details->running_rsc; lpc2 != NULL; lpc2 = lpc2->next) {
-                resource_t *rsc = (resource_t *) lpc2->data;
-
-                rsc->fns->print(rsc, "            ", print_opts | pe_print_rsconly, stream);
-            }
-            fprintf(stream, "        </node>\n");
-        } else {
-            fprintf(stream, "/>\n");
-        }
-    }
-    fprintf(stream, "    </nodes>\n");
-
-    /* Print resources section, if needed */
-    print_resources(stream, data_set, print_opts);
-
-    /* print Node Attributes section if requested */
-    if (show & mon_show_attributes) {
-        print_node_attributes(stream, data_set);
-    }
-
-    /* If requested, print resource operations (which includes failcounts)
-     * or just failcounts
-     */
-    if (show & (mon_show_operations | mon_show_failcounts)) {
-        print_node_summary(stream, data_set,
-                           ((show & mon_show_operations)? TRUE : FALSE));
-    }
-
-    /* If there were any failed actions, print them */
-    if (xml_has_children(data_set->failed)) {
-        print_failed_actions(stream, data_set);
-    }
-
-    /* Print stonith history */
-    if (fence_history) {
-        print_stonith_history(stdout, stonith_history);
-    }
-
-    /* Print tickets if requested */
-    if (show & mon_show_tickets) {
-        print_cluster_tickets(stream, data_set);
-    }
-
-    /* Print negative location constraints if requested */
-    if (show & mon_show_bans) {
-        print_neg_locations(stream, data_set);
-    }
-
-    fprintf(stream, "</crm_mon>\n");
-    fflush(stream);
-    fclose(stream);
 }
 
 /*!
@@ -3773,13 +1436,13 @@ print_html_status(pe_working_set_t * data_set,
         } else {
             fprintf(stream, "<font color=\"red\">OFFLINE</font>\n");
         }
-        if (print_brief && group_by_node) {
+        if (pp.print_brief && pp.group_by_node) {
             fprintf(stream, "<ul>\n");
             print_rscs_brief(node->details->running_rsc, NULL, print_opts | pe_print_rsconly,
                              stream, FALSE);
             fprintf(stream, "</ul>\n");
 
-        } else if (group_by_node) {
+        } else if (pp.group_by_node) {
             GListPtr lpc2 = NULL;
 
             fprintf(stream, "<ul>\n");
@@ -3801,16 +1464,16 @@ print_html_status(pe_working_set_t * data_set,
     print_resources(stream, data_set, print_opts);
 
     /* print Node Attributes section if requested */
-    if (show & mon_show_attributes) {
+    if (pp.show & mon_show_attributes) {
         print_node_attributes(stream, data_set);
     }
 
     /* If requested, print resource operations (which includes failcounts)
      * or just failcounts
      */
-    if (show & (mon_show_operations | mon_show_failcounts)) {
+    if (pp.show & (mon_show_operations | mon_show_failcounts)) {
         print_node_summary(stream, data_set,
-                           ((show & mon_show_operations)? TRUE : FALSE));
+                           ((pp.show & mon_show_operations)? TRUE : FALSE));
     }
 
     /* If there were any failed actions, print them */
@@ -3819,13 +1482,13 @@ print_html_status(pe_working_set_t * data_set,
     }
 
     /* Print failed stonith actions */
-    if (fence_history) {
+    if (pp.fence_history) {
         print_failed_stonith_actions(stream, stonith_history);
     }
 
     /* Print stonith history */
-    if (fence_history) {
-        if (show & mon_show_fence_history) {
+    if (pp.fence_history) {
+        if (pp.show & mon_show_fence_history) {
             print_stonith_history(stream, stonith_history);
         } else {
             print_stonith_pending(stdout, stonith_history);
@@ -3833,12 +1496,12 @@ print_html_status(pe_working_set_t * data_set,
     }
 
     /* Print tickets if requested */
-    if (show & mon_show_tickets) {
+    if (pp.show & mon_show_tickets) {
         print_cluster_tickets(stream, data_set);
     }
 
     /* Print negative location constraints if requested */
-    if (show & mon_show_bans) {
+    if (pp.show & mon_show_bans) {
         print_neg_locations(stream, data_set);
     }
 
@@ -4229,13 +1892,13 @@ mon_refresh_display(gpointer user_data)
 
     /* get the stonith-history if there is evidence we need it
      */
-    while (fence_history) {
+    while (pp.fence_history) {
         if (st != NULL) {
             if (st->cmds->history(st, st_opt_sync_call, NULL, &stonith_history, 120)) {
                 fprintf(stderr, "Critical: Unable to get stonith-history\n");
                 mon_cib_connection_destroy(NULL);
             } else {
-                if ((!fence_full_history) && (output_format != mon_output_xml)) {
+                if ((!pp.fence_full_history) && (output_format != mon_output_xml)) {
                     stonith_history = reduce_stonith_history(stonith_history);
                 }
                 stonith_history = sort_stonith_history(stonith_history);
@@ -4263,7 +1926,7 @@ mon_refresh_display(gpointer user_data)
     /* Unpack constraints if any section will need them
      * (tickets may be referenced in constraints but not granted yet,
      * and bans need negative location constraints) */
-    if (show & (mon_show_bans | mon_show_tickets)) {
+    if (pp.show & (mon_show_bans | mon_show_tickets)) {
         xmlNode *cib_constraints = get_object_root(XML_CIB_TAG_CONSTRAINTS,
                                                    mon_data_set->input);
         unpack_constraints(cib_constraints, mon_data_set);
@@ -4280,7 +1943,7 @@ mon_refresh_display(gpointer user_data)
             break;
 
         case mon_output_xml:
-            print_xml_status(mon_data_set, stonith_history);
+            print_xml_status(stdout, mon_data_set, stonith_history, pp);
             break;
 
         case mon_output_monitor:
