+++ tools/crm_simulate.c
@@ -570,26 +570,6 @@ profile_all(const char *dir, pe_working_set_t *data_set)
     }
 }
 
-static int
-count_resources(pe_working_set_t * data_set, resource_t * rsc)
-{
-    int count = 0;
-    GListPtr gIter = NULL;
-
-    if (rsc == NULL) {
-        gIter = data_set->resources;
-    } else if (rsc->children) {
-        gIter = rsc->children;
-    } else {
-        return is_not_set(rsc->flags, pe_rsc_orphan);
-    }
-
-    for (; gIter != NULL; gIter = gIter->next) {
-        count += count_resources(data_set, gIter->data);
-    }
-    return count;
-}
-
 int
 main(int argc, char **argv)
 {
